Fixes and smoothing the recent cherry-pick; making a smooth surface for Kokonusswasser.
