+++ nucypher/utilities/sandbox.py
@@ -97,8 +97,6 @@ def make_ursulas(ether_addresses: list,
                             federated_only=federated_only,
                             **ursula_kwargs)
 
-            ursula.attach_rest_server()
-
             class MockDatastoreThreadPool(object):
                 def callInThread(self, f, *args, **kwargs):
                     return f(*args, **kwargs)
@@ -121,7 +119,7 @@ def make_ursulas(ether_addresses: list,
             ursula.federated_only = True
 
         ursulas.add(ursula)
-        _TEST_KNOWN_URSULAS_CACHE[ursula.rest_interface.port] = ursula
+        _TEST_KNOWN_URSULAS_CACHE[ursula.rest_information()[0].port] = ursula
 
     if know_each_other and not bare:
 
@@ -192,6 +190,14 @@ class MockRestMiddleware(RestMiddleware):
     class NotEnoughMockUrsulas(MinerAgent.NotEnoughMiners):
         pass
 
+    def __get_mock_client_by_ursula(self, ursula):
+        port = ursula.rest_information()[0].port
+        return self.__get_mock_client_by_port(port)
+
+    def __get_mock_client_by_url(self, url):
+        port = int(url.split(":")[1])
+        return self.__get_mock_client_by_port(port)
+
     def __get_mock_client_by_port(self, port):  # TODO
         try:
             ursula = _TEST_KNOWN_URSULAS_CACHE[port]
@@ -203,25 +209,25 @@ class MockRestMiddleware(RestMiddleware):
         return mock_client
 
     def consider_arrangement(self, arrangement=None):
-        mock_client = self.__get_mock_client_by_port(arrangement.ursula.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(arrangement.ursula)
         response = mock_client.post("http://localhost/consider_arrangement", bytes(arrangement))
         assert response.status_code == 200
         return response
 
     def enact_policy(self, ursula, id, payload):
-        mock_client = self.__get_mock_client_by_port(ursula.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(ursula)
         response = mock_client.post('http://localhost/kFrag/{}'.format(id.hex()), payload)
         assert response.status_code == 200
         return True, ursula.stamp.as_umbral_pubkey()
 
     def send_work_order_payload_to_ursula(self, work_order):
-        mock_client = self.__get_mock_client_by_port(work_order.ursula.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(work_order.ursula)
         payload = work_order.payload()
         id_as_hex = work_order.arrangement_id.hex()
         return mock_client.post('http://localhost/kFrag/{}/reencrypt'.format(id_as_hex), payload)
 
     def get_treasure_map_from_node(self, node, map_id):
-        mock_client = self.__get_mock_client_by_port(node.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(node)
         return mock_client.get("http://localhost/treasure_map/{}".format(map_id))
 
     def node_information(self, host, port):
@@ -229,8 +235,8 @@ class MockRestMiddleware(RestMiddleware):
         response = mock_client.get("http://localhost/public_information")
         return response
 
-    def get_nodes_via_rest(self, address, port, announce_nodes=None, nodes_i_need=None):
-        mock_client = self.__get_mock_client_by_port(port)
+    def get_nodes_via_rest(self, url, announce_nodes=None, nodes_i_need=None):
+        mock_client = self.__get_mock_client_by_url(url)
 
         if nodes_i_need:
             # TODO: This needs to actually do something.
@@ -239,16 +245,16 @@ class MockRestMiddleware(RestMiddleware):
             pass
 
         if announce_nodes:
-            response = mock_client.post("https://{}:{}/node_metadata".format(address, port),
+            response = mock_client.post("https://{}/node_metadata".format(url),
                                      verify=False,
                                      data=bytes().join(bytes(n) for n in announce_nodes))  # TODO: TLS-only.
         else:
-            response = mock_client.get("https://{}:{}/node_metadata".format(address, port),
+            response = mock_client.get("https://{}/node_metadata".format(url),
                                     verify=False)  # TODO: TLS-only.
         return response
 
     def put_treasure_map_on_node(self, node, map_id, map_payload):
-        mock_client = self.__get_mock_client_by_port(node.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(node)
         response = mock_client.post("http://localhost/treasure_map/{}".format(map_id),
                       data=map_payload, verify=False)
         return response
